requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
beautifulsoup4==4.14.2
certifi==2025.11.12
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
fastapi==0.122.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.6.0
motor==3.7.1
packaging==25.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.4
pytest==9.0.1
python-dotenv==1.2.1
slowapi==0.1.9
soupsieve==2.8
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
wrapt==2.0.1
